refactor: category 위치 이동 #36
