+++ src/main/java/poomasi/domain/product/_category/repository/CategoryRepository.java
@@ -1,7 +1,7 @@
-package poomasi.domain.category.repository;
+package poomasi.domain.product._category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import poomasi.domain.category.entity.Category;
+import poomasi.domain.product._category.entity.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
