+++ src/main/java/poomasi/domain/product/_category/service/CategoryService.java
@@ -1,11 +1,11 @@
-package poomasi.domain.category.service;
+package poomasi.domain.product._category.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import poomasi.domain.category.dto.CategoryResponse;
-import poomasi.domain.category.entity.Category;
-import poomasi.domain.category.repository.CategoryRepository;
+import poomasi.domain.product._category.dto.CategoryResponse;
+import poomasi.domain.product._category.entity.Category;
+import poomasi.domain.product._category.repository.CategoryRepository;
 
 @Service
 @RequiredArgsConstructor
